requirements.txt
openai>=1.0.0
watchdog==3.0.0
python-dotenv>=1.0.0
pydub>=0.25.1
pyinstaller==6.4.0
tkinter 